Move mobile open orders table into main
